Using only one connection, instead of creating multiple. Fixed `Cannot use a session that has ended`
