+++ bot.js
@@ -12,14 +12,16 @@ const logger = require('./util/logger');
 const loadCommands = require('./commands/loadCommands');
 const loadFeatures = require('./features/loadFeatures');
 const messageCounter = require('./util/messageCount');
+const mongo = require('./util/mongo');
 
 // commands
 client.on('ready', async () => {
   console.log('The client is ready!');
+
+  await mongo();
   // Load commands and features
   loadCommands(client);
   loadFeatures(client);
-  messageCounter(client);
 });
 /*
 client.on('ready', () => {
