+++ util/mongo.js
@@ -5,6 +5,7 @@ const { mongoPath } = require('../token.json');
 // connects to server
 module.exports = async () => {
   await mongoose.connect(mongoPath, {
+    keepAlive: true,
     useFindAndModify: false,
     useNewUrlParser: true,
     useUnifiedTopology: true,
